fix: address issue #490 (Ashigaru autonomous run)
